refactor: restore lib headroom on overlay, map, trade, and victory files (#4642)
Topic-split the seven near-cap app/lib modules so each resulting file is ≤230 lines without renaming overlay/screen hosts or UiScreenIds.
