+++ app/lib/features/game/flame/map_area/map_area.dart
@@ -1,4 +1,5 @@
 export 'game_map_area_background.dart';
 export 'game_map_canvas_stack.dart';
+export 'game_map_canvas_stack_selection_prompt_tokens.dart';
 export 'game_map_canvas_stack_hover.dart';
 export 'game_map_canvas_stack_selection_prompt.dart';
