+++ app/lib/features/game/widgets/province_overlay/province_sea_zone_detail_overlay_tile_section_tokens.dart
@@ -1,0 +1,2 @@
+/// Disabled-icon alpha for province overlay tile inline [CtIconAction]s.
+const double kProvinceOverlayTileInlineActionDisabledAlpha = 0.65;
